+++ lib/src/constants.dart
@@ -14,7 +14,7 @@ class ConstHelper {
   static const String replaceIfFirstLetterNumber = 'n';
 
   static const String configName = 'turn_gen';
-  static const String version = '0.3.0';
+  static const String version = '0.3.2';
 }
 
 class ConstConsole {