+++ lib/src/enum_v2/models/enum_v2_common.dart
@@ -1,0 +1,80 @@
+// ignore_for_file: sort_constructors_first, avoid_bool_literals_in_conditional_expressions
+
+import 'package:turn_gen/src/common/common.dart';
+import 'package:turn_gen/src/enum_v2/enum_v2.dart';
+
+/* no: hash tojson fromjson*/
+class EnumV2CommonModel { 
+  /* init: '' */
+  final String nameClass;
+  /* init: EnumTypeVarable.none*/
+  final EnumTypeVarable typeEnum;
+  /* init: '' */
+  final String nameValue;
+  /* init: '' */
+  final String headerClass;
+
+  /* init: '' */
+  final String contentToEnd;
+  /* init: false */
+  final bool isDefault;
+    /* init: false */
+  final bool isCanNull;
+  /* init: const []*/
+  final List<EnumV2ItemModel> listItem;
+
+  // end
+   
+//          --TURN_GEN--
+//             v0.3.2
+//  *************************************
+//           GENERATED CODE 
+//  *************************************
+  
+  const EnumV2CommonModel({
+    this.nameClass = '',
+    this.typeEnum = EnumTypeVarable.none,
+    this.nameValue = '',
+    this.headerClass = '',
+    this.contentToEnd = '',
+    this.isDefault = false,
+    this.isCanNull = false,
+    this.listItem = const [],
+  });
+  /*
+   factory EnumV2CommonModel.init() => EnumV2CommonModel(
+      ); 
+  */
+
+
+  EnumV2CommonModel copyWith({
+    String? nameClass,
+    EnumTypeVarable? typeEnum,
+    String? nameValue,
+    String? headerClass,
+    String? contentToEnd,
+    bool? isDefault,
+    bool? isCanNull,
+    List<EnumV2ItemModel>? listItem,
+  }) {
+    return EnumV2CommonModel(
+      nameClass: nameClass ?? this.nameClass, 
+      typeEnum: typeEnum ?? this.typeEnum, 
+      nameValue: nameValue ?? this.nameValue, 
+      headerClass: headerClass ?? this.headerClass, 
+      contentToEnd: contentToEnd ?? this.contentToEnd, 
+      isDefault: isDefault ?? this.isDefault, 
+      isCanNull: isCanNull ?? this.isCanNull, 
+      listItem: listItem ?? this.listItem, 
+    );
+  }
+
+  
+  
+
+      @override
+  String toString() {
+    return 'EnumV2CommonModel(nameClass: $nameClass, typeEnum: $typeEnum, nameValue: $nameValue, headerClass: $headerClass, contentToEnd: $contentToEnd, isDefault: $isDefault, isCanNull: $isCanNull, listItem: $listItem, )';
+    }
+
+}