+++ lib/src/union_class/model/union_item_model.dart
@@ -1,10 +1,25 @@
+// ignore_for_file: sort_constructors_first, avoid_bool_literals_in_conditional_expressions
 import 'package:turn_gen/src/src.dart';
 
 /* no: hash tojson fromjson*/
 class UnionItemModel {
+  /* init: '' */
+  final String nameUnion;
+  /* init: '' */
+  final String paramStr;
+  /* init: '' */
+  final String comment;
+  /* init: const {} */
+  final Map<String, String> mapNameWithTag;
+  /*   init: const[]    */
+  final List<UnionParameterModel> listParameters;
+  /*  init: EnumParameter.none   */
+  final EnumParameter parameter;
+
   // end
 
 //          --TURN_GEN--
+//             v0.3.0
 //  *************************************
 //           GENERATED CODE
 //  *************************************
@@ -12,22 +27,11 @@ class UnionItemModel {
   const UnionItemModel({
     this.nameUnion = '',
     this.paramStr = '',
+    this.comment = '',
     this.mapNameWithTag = const {},
     this.listParameters = const [],
     this.parameter = EnumParameter.none,
   });
-
-  /* init: '' */
-  final String nameUnion;
-  /* init: '' */
-  final String paramStr;
-
-  /* init: const {} */
-  final Map<String, String> mapNameWithTag;
-  /*   type: List<data>   init: const[]    */
-  final List<UnionParameterModel> listParameters;
-  /* type: enum   init: EnumParameter.none   */
-  final EnumParameter parameter;
   /*
    factory UnionItemModel.init() => UnionItemModel(
       ); 
@@ -36,6 +40,7 @@ class UnionItemModel {
   UnionItemModel copyWith({
     String? nameUnion,
     String? paramStr,
+    String? comment,
     Map<String, String>? mapNameWithTag,
     List<UnionParameterModel>? listParameters,
     EnumParameter? parameter,
@@ -43,6 +48,7 @@ class UnionItemModel {
     return UnionItemModel(
       nameUnion: nameUnion ?? this.nameUnion,
       paramStr: paramStr ?? this.paramStr,
+      comment: comment ?? this.comment,
       mapNameWithTag: mapNameWithTag ?? this.mapNameWithTag,
       listParameters: listParameters ?? this.listParameters,
       parameter: parameter ?? this.parameter,
@@ -51,6 +57,6 @@ class UnionItemModel {
 
   @override
   String toString() {
-    return 'UnionItemModel(nameUnion: $nameUnion, paramStr: $paramStr, mapNameWithTag: $mapNameWithTag, listParameters: $listParameters, parameter: $parameter, )';
+    return 'UnionItemModel(nameUnion: $nameUnion, paramStr: $paramStr, comment: $comment, mapNameWithTag: $mapNameWithTag, listParameters: $listParameters, parameter: $parameter, )';
   }
 }